adjusting UI, changing batt icon to voltage based
I've updated the code to calculate the battery icon from the voltage for the time being.
if/when the RFID component gets off the ground, and I can successfully store the usage, I'll change it back to use that.

I've also made a few tweaks to the UI, added an animated startup logo, and  fixed an icon in battery.h that wasn't in the correct format.